Mouse screen pos

+++ Engine/Code/Engine/Input/InputUtils.h
@@ -7,7 +7,7 @@
 //-------------------------------------------------------------------------------------
 // SUPPORTED KEY CODES -- based on windows VKs
 //
-enum class KeyCode : uint8_t
+enum class KeyCode : int
 {
     F1 = 112, F2 = 113, F3 = 114, F4 = 115, F5 = 116, F6 = 117, 
     F7 = 118, F8 = 119, F9 = 120, F10 = 121, F11 = 122,
